AP_ICEngine: Change throttle_override to pass in current throttle value

+++ libraries/AP_ICEngine/AP_ICEngine.h
@@ -31,7 +31,7 @@ public:
     void update(void);
 
     // check for throttle override
-    bool throttle_override(uint8_t &percent);
+    bool throttle_override(float &percent);
 
     enum ICE_State {
         ICE_OFF=0,
